Full custom palette is no longer required
Changed --clut to accept a partial palette, which forms the start
of the final CLUT. This allows the index of some colours to be
fixed, which is useful for dynamic colours.

Any additional colours required by the image are appended to the
manual list. The final palette size must still fit the mode.