Fixed odd width assumption that lost right-most pixels
We previously assumed that the sprite foreground would be an odd width so the
shifted version would still be within the same display byte boundary. If the
right-most pixel was used it would be lost instead of spilling into the next
display byte.

We now generate separate routines for even and odd shifts, but only if needed
for the specific sprite data being used. Even widths will generate more code,
unless --shift=0 is used to specify only even positioning.