Share common even/odd code header
If the even/odd routines start with common code we can put it before the
branch point, but only if it doesn't affect the carry flag. Another exception
is stack saving that needs to self-modify a routine-specific variable.

If either routine is reduced to just the trailing "ret" we can conditionally
return instead, leaving the alternative code to run without a jump. If a jump
is needed and the 'even' code is short enough we'll use a relative jump
instead of absolute to save space.