Add --share flag to share save/restore code
Sprites with even widths using the full width require an extra byte when
shifted. This requires unique save/restore code, which increases code size.

This new flag forces a single version of the code covering the combined largest
width for even/odd cases. This reduces code size but comes with a risk that it
can't be used with the sprite flush against the right edge of the display,
without overflowing the screen row. Use with care!