Added new 'copy' routine for sprite restoring
This routine expects a pristine copy of the display to be available in the
opposite 32K from the drawn display. It undraws a sprite by copying clean
versions of any display bytes used by the sprite.

Performance is around the same as the 'save' routine, giving a faster option to
restore under masked sprites.