Improved support for background transparency
Previously we assumed that palette position 0 was always black and used as a
transparent background colour. This prevented uses of black as opaque.

A new --bkgcol option allows the background colour to be set to any SAM palette
index. This allows sprite sheets to use a different colour for the background.

If no background colour is specified any palette colours in the source image
with zero alpha are treated as background. If none are found we fall back on
using palette 0 as black background like before.

Also changed to use the PIL quantize function to map to SAM colours, which
performs the same Euclidian distance calculation as the original code. This
didn't work in earlier versions but now seems reliable.